use Junit files instead of csv
Create the test report from Junit files instead of csv files.
This adds the support for test name, package, classname the number of
failed tests and number of tests.

Add the possibility to add a Test ID column using the '-s' argument.
In order to generate this column, test name should include the ID
separate from the rest with ' - '.

Change-Id: I87fcb07c298cd5905b07ba8dd342fd0bda157f83