compile.sh: bug correction when suite is default
When no suite is logged in cukinia.conf the suite is named "default".
In that case, instead of printing the suite name, the script prints the
xml file name.

This commit corrects a bug where no tests where present at all in the pdf
when the suite was the default one.

Change-Id: I8c7265c0f939b54a84929449e2493ff4e3e47463