compile.sh: correct bug of last test not in report
It was noticed that the last test of each category wasn't displayed in
the report. The problem was obviously a mistake in a loop that doesn't
handle the last test.

In fact, the while loop reading test cases uses bash process substitution
which is a bad idea. This method split lines on newline character and
will thus not handle the last line if it is not terminated by a newline
character.

Piping a variable directly in the while loop do this job and the last
test appear.

Change-Id: I60f36e615a46eb1e24e85aa59b172c7617b3de9d