compile.sh: add basic display for compliance matrix
One or many compliance matrix will be generated by compile.sh and added
to the adoc.
For now on, all csv files in the source directory are interpreted as
compliance matrix.
This commit just adds mechanism to read the csv file that contains the
matrix information and display a basic table.

SFL: #8700
Change-Id: I90e4edddf64c7438a17aee949f7a84f51d9d666f