compile.sh: store file names in an array
Test file names are found at the beginning of the script and reuse in
the compliance matrix file. So they need to be stored in a variable.

To avoid word splitting problems, if a file contains a space, each name
has to be enclosed with double quotes separately.
The only solution to do that is to use bash arrays.

The syntax is much heavier, so I choose to put TEST_FILES variable as
global instead of passing it to add_compliance_matrix function.

SFL: #8700
Change-Id: I9d0b004dad00a1d459faa276445bd53515cdd7db