compile.py: add support of cukinia id
Previously, the creation of the report was getting the test id in the
test description if the split_test_id argument was used. Now that
Cukinia is supporting test id, the report generation is checking for an
id in the first test. When an id is detected, the report generation is
adding a Test id column and matrix compliance when provided.

SFL: #9523
Change-Id: I040f83570cc92fac38d249407da98f8f205dd483