compile.py: create one compliance matrix per machine
When the option `-m` is active, one matrix is created per machine. This
matrix only looks for the test results on its specific machine.
When the option is not there, only one matrix is created. It looks for
the test results on all machines at once.

This commit also update the examples, to avoid absent test. Otherwise,
the CI will fail.

Change-Id: I44bc29d1554f225b40f022fe2a9f1be763e21e15