compile.py: update check_for_id behavior
The check_for_id function only checked the first test to see if the test
ID was used. Therefore, if one of the tests didn't use an ID, the
behavior was different depending on whether it was the first in the
suite or not.

The test ID column is now added as soon as a test of a suite has an ID.