compile.py: enable compliance matrix if test id is present in one suite
In case test id was not enabled in the last file or the last test suite
parsed by compile.py, the compliance matrix was not included, and the
script was failing. Then change the behavior to enable the compliance
matrix if at least one test suite has the test id feature enabled.

SFL: #17866
Change-Id: Id34b20af710355be98d65c2d749a24d80ded812a